requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
supabase==2.2.0
gotrue==2.8.0
Werkzeug==2.3.7
boto3==1.28.57
psycopg2-binary==2.9.7
httpx==0.24.1
requests==2.32.5
apscheduler

